+++ mockspresso-mockito/src/test/java/com/episode6/hackit/mockspresso/mockito/integration/MockitoKotlinExtensionTest.kt
@@ -4,8 +4,13 @@ import com.episode6.hackit.mockspresso.BuildMockspresso
 import com.episode6.hackit.mockspresso.annotation.Dependency
 import com.episode6.hackit.mockspresso.basic.plugin.simple.SimpleInjectMockspressoPlugin
 import com.episode6.hackit.mockspresso.dependencyOf
+import com.episode6.hackit.mockspresso.mockito.Conditions.mockCondition
 import com.episode6.hackit.mockspresso.mockito.MockitoPlugin
+import com.episode6.hackit.mockspresso.realClassOf
+import com.episode6.hackit.mockspresso.realImplOf
 import com.episode6.hackit.mockspresso.reflect.NamedAnnotationLiteral
+import com.episode6.hackit.mockspresso.testing.isConcreteInstanceOf
+import com.episode6.hackit.mockspresso.testing.satisfies
 import org.fest.assertions.api.Assertions.assertThat
 import org.junit.Rule
 import org.junit.Test
@@ -16,14 +21,20 @@ import javax.inject.Named
  */
 class MockitoKotlinExtensionTest {
 
-  private interface TestInterface
-  private class TestDependency : TestInterface
-  private class TestObjectWithConcrete(val testDep: TestDependency)
-  private class TestObjectWithInterface(val testDep: TestInterface)
-  private class TestObjectWithAnnotatedInterface(@Named("testing") val testDep: TestInterface)
+  private interface TestDependencyInterface
+  private interface TestObjectInterface
+  private class TestDependency : TestDependencyInterface
+
+  private class TestObjectWithConcrete(val testDep: TestDependency) : TestObjectInterface
+  private class TestObjectWithInterface(val testDep: TestDependencyInterface)
+  private class TestObjectWithAnnotatedInterface(@Named("testing") val testDep: TestDependencyInterface)
   private class TestResources {
-    @Dependency(bindAs = TestInterface::class) @field:Named("testing") val testDependency = TestDependency()
+    @Dependency(bindAs = TestDependencyInterface::class) @field:Named("testing") val testDependency = TestDependency()
   }
+
+  private class OuterTestObjectWithConcreteObj(val testObj: TestObjectWithConcrete)
+  private class OuterTestObjectWithInterface(val testObj: TestObjectInterface)
+  private class OuterTestObjectWithAnnotatedInterface(@Named("testing")  val testObj: TestObjectInterface)
 
   @get:Rule val mockspresso = BuildMockspresso.with()
       .plugin(SimpleInjectMockspressoPlugin())
@@ -43,7 +54,7 @@ class MockitoKotlinExtensionTest {
 
   @Test fun testInterfaceDependency() {
     val testObject: TestObjectWithInterface = mockspresso.buildUpon()
-        .dependencyOf<TestInterface> { testDependency }
+        .dependencyOf<TestDependencyInterface> { testDependency }
         .build()
         .create(TestObjectWithInterface::class.java)
 
@@ -54,7 +65,7 @@ class MockitoKotlinExtensionTest {
     // we can't define an annotation literal in kotlin, but we can use one
     // that is defined in java
     val testObject: TestObjectWithAnnotatedInterface = mockspresso.buildUpon()
-        .dependencyOf<TestInterface>(qualifier = NamedAnnotationLiteral("testing")) { testDependency }
+        .dependencyOf<TestDependencyInterface>(qualifier = NamedAnnotationLiteral("testing")) { testDependency }
         .build()
         .create(TestObjectWithAnnotatedInterface::class.java)
 
@@ -71,5 +82,50 @@ class MockitoKotlinExtensionTest {
         .create(TestObjectWithAnnotatedInterface::class.java)
 
     assertThat(testObject.testDep).isEqualTo(testResources.testDependency)
+  }
+
+  @Test fun testRealClassOf() {
+    val outerObj: OuterTestObjectWithConcreteObj = mockspresso.buildUpon()
+        .realClassOf<TestObjectWithConcrete>()
+        .dependencyOf { testDependency }
+        .build()
+        .create(OuterTestObjectWithConcreteObj::class.java)
+
+    assertThat(outerObj.testObj)
+        .isNot(mockCondition())
+        .isConcreteInstanceOf<TestObjectWithConcrete>()
+        .satisfies {
+          assertThat(it.testDep).isEqualTo(testDependency)
+        }
+  }
+
+  @Test fun testRealImplOf() {
+    val outerObj: OuterTestObjectWithInterface = mockspresso.buildUpon()
+        .realImplOf<TestObjectInterface, TestObjectWithConcrete>()
+        .dependencyOf { testDependency }
+        .build()
+        .create(OuterTestObjectWithInterface::class.java)
+
+    assertThat(outerObj.testObj)
+        .isNot(mockCondition())
+        .isConcreteInstanceOf<TestObjectWithConcrete>()
+        .satisfies {
+          assertThat(it.testDep).isEqualTo(testDependency)
+        }
+  }
+
+  @Test fun testAnnotatedRealImplOf() {
+    val outerObj: OuterTestObjectWithAnnotatedInterface = mockspresso.buildUpon()
+        .realImplOf<TestObjectInterface, TestObjectWithConcrete>(qualifier = NamedAnnotationLiteral("testing"))
+        .dependencyOf { testDependency }
+        .build()
+        .create(OuterTestObjectWithAnnotatedInterface::class.java)
+
+    assertThat(outerObj.testObj)
+        .isNot(mockCondition())
+        .isConcreteInstanceOf<TestObjectWithConcrete>()
+        .satisfies {
+          assertThat(it.testDep).isEqualTo(testDependency)
+        }
   }
 }